+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 ==========
 - Modernize project definition to latest Python best practices. Thanks, @surister.
 - Exceptions: Exceptions from the BLOB API now include their full names.
+- Added JWT token authentication
 
 2025/01/30 2.0.0
 ================